==> code/10_appendix_table_a2.Rmd <==
---
title: "Untitled"
author: "Hunter York"
date: "2023-03-09"
output: 
  pdf_document:
    keep_tex: true
geometry: paperheight=35cm,paperwidth=25cm,margin=0.2cm
header-includes:
    - \usepackage{caption}
    - \usepackage{booktabs}
    - \AtBeginDocument{\let\maketitle\relax}
    - \captionsetup[table]{labelformat=empty}
    - \pagenumbering{gobble}
---



```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = F)
library(data.table)
library(kableExtra)
library(knitr)
library(dplyr)
library(readxl)
```

```{r, results = 'asis'}
table1 <- fread("../ref/xwalks_occ_soc.csv")
table1[, OCCSOC := paste0(substr(OCCSOC,1,2), "", substr(OCCSOC, 3, 7), "")]
table1[, OCC1950 := as.numeric(OCC1950)]
table1 <- table1[order(OCC1950)]

education_2018 <- read_excel("../ref/db_22_2_excel 2018.2/Education, Training, and Experience.xlsx") %>% data.table()
temp <- education_2018[,.(`O*NET-SOC Code`, Title)] %>% unique()
temp[, OCCSOC := substr(`O*NET-SOC Code`,1,7)]
temp <- temp[, Title := Title[1], by = OCCSOC]
temp <- temp[,.(OCCSOC, Title)] %>% unique()

table1 <- merge(table1, temp[,.(OCCSOC, Title)])
table1 <- table1[order(OCC1950)]
table1 <- table1[,.(OCC1950, OCC1950_name, OCCSOC, Title)]


table1 %>%  # mutate_all(linebreak, double_escape = F, linebreaker = "\\n") %>% 
  kable("latex", row.names = F, escape = T,align = c("r", rep("r", 4)), #booktabs = T,
     #   caption = "Table 1: Occupational Structure",
        booktabs = TRUE,
        longtable = TRUE, col.names = c( "Occ. Code", "1950 Census Occupation Title", "SOC Code", "SOC Title")) %>%
  kable_styling(., latex_options = c("hold_position", "repeat_header")) %>% 
   column_spec(c(1,3), width = "2cm") %>% 
  column_spec(c(2, 4), width = "8cm")


```
