==> code/13_appendix_table_a5.Rmd <==
---
title: "table_appendix_4"
author: "Hunter York"
date: "2023-04-19"
output:
  pdf_document:
    keep_tex: true
geometry: paperheight=8in,paperwidth=13.3in,margin=0.2in
header-includes:
    - \usepackage{booktabs}
    - \usepackage{dcolumn}
    - \AtBeginDocument{\let\maketitle\relax}
    - \usepackage[labelformat=empty]{caption}
    - \pagenumbering{gobble} 
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = F, warning = F, message = F)
library(data.table)
library(dplyr)


model_version <- "20240123_first_Resub_trial"

```
```{r, results = 'asis'}
table2 <- readRDS( paste0("../outputs/", model_version, "/table2.rds"))

setnames(table2, c("ll", "aic", "bic", "delta", "df", "q","df.x", "Chisq.x", 
                   "Pr(>Chisq).x", "hm.x", "df.y", "Chisq.y", 
                   "Pr(>Chisq).y","hm.y", "df.z", "Chisq.z", 
                   "Pr(>Chisq).z","hm.z", 
                   "df.a", "Chisq.a", 
                   "Pr(>Chisq).a","hm.a", "sex", "parent_sex" ))
table2[, Chisq.x  := round(Chisq.x)]
table2[, Chisq.y  := round(Chisq.y)]
table2[, Chisq.z  := round(Chisq.z)]
table2[, Chisq.a  := round(Chisq.a)]
table2[, ll  := round(ll)]

table2[ ,`Pr(>Chisq).x` := ifelse(`Pr(>Chisq).x` < .001,"***", 
                                  ifelse(`Pr(>Chisq).x` < .01, "**", 
                                         ifelse(`Pr(>Chisq).x` < .05, "*", 
                                                ifelse(`Pr(>Chisq).x` >=.05, "NS", NA))))]
table2[ ,`Pr(>Chisq).y` := ifelse(`Pr(>Chisq).y` < .001,"***", 
                                  ifelse(`Pr(>Chisq).y` < .01, "**", 
                                         ifelse(`Pr(>Chisq).y` < .05, "*", 
                                                ifelse(`Pr(>Chisq).y` >=.05, "NS", NA))))]
table2[ ,`Pr(>Chisq).z` := ifelse(`Pr(>Chisq).z` < .001,"***", 
                                  ifelse(`Pr(>Chisq).z` < .01, "**", 
                                         ifelse(`Pr(>Chisq).z` < .05, "*", 
                                                ifelse(`Pr(>Chisq).z` >=.05, "NS", NA))))]
table2[ ,`Pr(>Chisq).a` := ifelse(`Pr(>Chisq).a` < .001,"***", 
                                  ifelse(`Pr(>Chisq).a` < .01, "**", 
                                         ifelse(`Pr(>Chisq).a` < .05, "*", 
                                                ifelse(`Pr(>Chisq).a` >=.05, "NS", NA))))]
table2[is.na(table2)] <- ""

library(kableExtra)
table2 <- table2[,.( ll, bic, df, df.x, Chisq.x,
                     `Pr(>Chisq).x`, df.y, Chisq.y,
                     `Pr(>Chisq).y`,  df.z, Chisq.z,
                     `Pr(>Chisq).z`,  df.a, Chisq.a, `Pr(>Chisq).a`)]



table2[c(1,9,17,25), model :=     "$\\;$Model 1. $\\:$Base Model"]
table2[c(1,9,17,25) + 1, model := "$\\;$Model 2. $\\:\\quad$Model 1 + Macro-/Meso-/Micro- Class Immobility"]
table2[c(1,9,17,25) + 2, model := "$\\;$Model 3. $\\:\\quad\\quad$Model 2 + Meso-Class Mobility"]
table2[c(1,9,17,25) + 3, model := "$\\;$Model 4. $\\:\\quad$Model 1 + SEI"]
table2[c(1,9,17,25) + 4, model := "$\\;$Model 5. $\\:\\quad\\quad$Model 4 + Macro-/Meso-/Micro- Class Immobility"]
table2[c(1,9,17,25) + 5, model := "$\\;$Model 6. $\\:\\quad\\quad\\quad$Model 5 + 10 Gradational Effects"]
#table2[c(1,9,17,25) + 4, model := "$\\;$Model 5. $\\:\\quad\\quad\\quad$Model 3 + 4 Gradational RC Effects"]
table2[c(1,9,17,25) + 6, model := "$\\;$Model 7. $\\:\\quad$Model 1 + 10 Gradational Effects"]
table2[c(1,9,17,25) + 7, model := "$\\;$Model 8. $\\:\\quad\\quad$Model 7 + SEI"]




table2 <- table2[,.(model, ll, bic, df, df.x, Chisq.x, `Pr(>Chisq).x`, df.y, Chisq.y, `Pr(>Chisq).y`,  df.z, Chisq.z, `Pr(>Chisq).z`,  df.a, Chisq.a, `Pr(>Chisq).a`)]

nums <- names(table2)[names(table2) %like% "aic|bic|df|^Chisq|ll"]
table2[,(nums) := lapply(.SD, as.numeric), .SDcols = nums]

table2[,(nums) := lapply(.SD, formatC,  big.mark = ",", format = "d"), .SDcols = nums]
table2[is.na(table2)] <- ""
table2[table2 == "NA"] <- ""

knitr::kable(table2,"latex", row.names = F,longtable = T, escape = F,align = c("l", rep("r", 5),
                                                                 "c", rep("r", 2), "c", rep("r", 2), "c",  rep("r", 2), "c"),
             col.names = c( "Model", "$LL$","$BIC$", "$df$","$\\Delta df$", "Chi Sq.", "P-Value", "$\\Delta df$", "Chi Sq.", "P-Value","$\\Delta df$", "Chi Sq.", "P-Value", "$\\Delta df$", "Chi Sq.", "P-Value"), 
             booktabs = TRUE,
             linesep = "")%>%
  # kableExtra::row_spec(c(1,3,5,7,15,17,19, 21), background = "#D0D0D0") %>% #darkest
  # kableExtra::row_spec(c(c(1,3,5,7,15,17,19)+1),
  #                      background = "#E8E8E8") %>% #medium dark
  # kableExtra::row_spec(c(8,10,12,14, 22,24,26, 28), background = "#F8F8F8") %>% #light
  
  # kable_styling("striped", full_width = F,
  #               position = "left", font_size = 12) %>%
  add_header_above(c(" " =4, "Model 1 LR Test" = 3,  "Model 4 LR Test" = 3,  "Model 5 LR Test" = 3, "Model 7 LR Test" = 3)) %>% 
  # add_footnote(c("LR Test denotes likelihood-ratio tests", 
  #                "* : p < .05, ** : p < .01, *** : p < .001"),
  #              notation = "number", escape = F)%>% 
  group_rows(group_label = "Fathers/Sons", 
             start_row = 1, end_row = 8, 
             latex_align = "l", 
             latex_wrap_text = T, 
             bold = T) %>% 
  group_rows(group_label = "Fathers/Daughters", 
             start_row = 9, end_row = 16, 
             latex_align = "l", 
             latex_wrap_text = T, 
             bold = T) %>% 
  group_rows(group_label = "Mothers/Sons", 
             start_row = 17, end_row = 24, 
             latex_align = "l", 
             latex_wrap_text = T, 
             bold = T) %>% 
  group_rows(group_label = "Mothers/Daughters", 
             start_row = 25, end_row = 32, 
             latex_align = "l", 
             latex_wrap_text = T, 
             bold = T) %>% 
  kable_styling(font_size = 7) %>%
  column_spec(1, width = "85mm") %>% 
    column_spec(c(2:4, 6:7, 9:10, 12:13, 15:16), width = "10mm") %>% 
    column_spec(c(5, 8, 11, 14), width = "6mm") %>% 

  save_kable(file = paste0("../outputs/", model_version, "/table_appendix_4.tex"))
```
